Rename Service Worker initialization module file
